Rename types, move types to popover-item folder
